before qu gu ren

+++ include/hypertea/util/device_alternate.hpp
@@ -109,7 +109,7 @@ namespace hypertea {
 const size_t HYPERTEA_OPENCL_NUM_THREADS = 128;
 
 // OPENCL: number of blocks for threads.
-inline int HYPERTEA_GET_BLOCKS(const int N) {
+inline size_t HYPERTEA_GET_BLOCKS(const size_t N) {
   return (N + HYPERTEA_OPENCL_NUM_THREADS - 1) / HYPERTEA_OPENCL_NUM_THREADS * HYPERTEA_OPENCL_NUM_THREADS;
 }
